# faster by preventing string copy