before:
```
name ............................... min time      avg time    std dv   runs
findAll ............................ 0.025 ms      0.027 ms    ±0.002  x1000
findAll ** ......................... 0.005 ms      0.006 ms    ±0.000  x1000
findAll **/.. ...................... 0.268 ms      0.285 ms    ±0.014  x1000
findAll */.. ....................... 0.024 ms      0.027 ms    ±0.002  x1000
findAll ../*/.. .................... 0.025 ms      0.028 ms    ±0.002  x1000
```
after:
```
name ............................... min time      avg time    std dv   runs
findAll ............................ 0.016 ms      0.017 ms    ±0.001  x1000
findAll ** ......................... 0.006 ms      0.006 ms    ±0.000  x1000
findAll **/.. ...................... 0.172 ms      0.176 ms    ±0.005  x1000
findAll */.. ....................... 0.016 ms      0.018 ms    ±0.001  x1000
findAll ../*/.. .................... 0.017 ms      0.019 ms    ±0.002  x1000
```